refactor(api): split routes_shared and import_project
