+++ tools/size_baseline.txt
@@ -14,6 +14,7 @@ src/quodeq/api/_evaluation_routes.py:130:function
 src/quodeq/api/_evaluation_routes.py:183:function
 src/quodeq/api/_evaluation_routes.py:1:file
 src/quodeq/api/_evaluation_routes.py:251:function
+src/quodeq/api/_import_validation.py:88:function
 src/quodeq/api/_log_stream_routes.py:110:function
 src/quodeq/api/_log_stream_routes.py:135:function
 src/quodeq/api/_run_event_stream.py:1:file
@@ -24,18 +25,17 @@ src/quodeq/api/assistant_routes.py:1:file
 src/quodeq/api/assistant_routes.py:239:function
 src/quodeq/api/assistant_routes.py:354:function
 src/quodeq/api/assistant_workspace_routes.py:25:function
-src/quodeq/api/import_project.py:131:function
-src/quodeq/api/import_project.py:1:file
-src/quodeq/api/import_project.py:362:function
+src/quodeq/api/import_project.py:179:function
 src/quodeq/api/llm_bridge_routes.py:54:function
 src/quodeq/api/routes_findings.py:90:function
 src/quodeq/api/routes_project_data.py:15:function
 src/quodeq/api/routes_project_list.py:1:file
 src/quodeq/api/routes_project_list.py:235:function
 src/quodeq/api/routes_project_list.py:66:function
-src/quodeq/api/routes_shared.py:142:function
-src/quodeq/api/routes_shared.py:1:file
-src/quodeq/api/routes_shared.py:275:function
+src/quodeq/api/routes_shared_config.py:27:function
+src/quodeq/api/routes_shared_mirrors.py:41:function
+src/quodeq/api/routes_shared_pull.py:27:function
+src/quodeq/api/routes_shared_pull.py:30:function
 src/quodeq/api/standards_import_routes.py:15:function
 src/quodeq/api/standards_overrides_routes.py:29:function
 src/quodeq/api/standards_visibility_routes.py:30:function
